+++ src/lib/domain/services/state_service.ts
@@ -7,9 +7,9 @@ import { calcMaxHP } from '$lib/domain/services/attribute_service';
 import { buildPlayerFromCsv, buildEnemyFromCsv } from '$lib/data/repositories/character_repository';
 import { randomName } from '$lib/data/repositories/random_name_repository';
 import { performAction } from '$lib/domain/services/action_executor';
-import { waitForAnimationsComplete } from '$lib/presentation/utils/effectBus';
+import { waitForAnimationsComplete } from '$lib/presentation/utils/effect_bus';
 import { randomEvent } from '$lib/domain/services/event_service';
-import { pushLog, setLogState, resetDisplayLogs } from '$lib/presentation/utils/logUtil';
+import { pushLog, setLogState, resetDisplayLogs } from '$lib/presentation/utils/log_util';
 import { getRewardsForEnemy } from '$lib/data/repositories/reward_repository';
 import { tickStatusesTurnStart } from '$lib/data/consts/statuses';
 import { createCompanionRepository } from '$lib/data/repositories/companion_repository';
